Authorization Commit

+++ app/Http/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Comment;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -18,6 +19,7 @@ class CommentController extends Controller
         $comment = new Comment;
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -25,8 +27,12 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back();
+        if( Gate::allows('comment-delete', $comment) ) {
+            $comment->delete();
+            return back();
+        } else {
+            return back()->with('error', 'Unauthorize');
+        }
     }
 }
